// app/src/generated/models/propertyResponsePropertyCategory.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * High-level property classification
 */
export type PropertyResponsePropertyCategory =
  (typeof PropertyResponsePropertyCategory)[keyof typeof PropertyResponsePropertyCategory];

export const PropertyResponsePropertyCategory = {
  RESIDENTIAL: 'RESIDENTIAL',
  COMMERCIAL: 'COMMERCIAL',
  INDUSTRIAL: 'INDUSTRIAL',
  AGRICULTURAL: 'AGRICULTURAL',
  MIXED_USE: 'MIXED_USE',
} as const;