// app/src/generated/models/propertyResponsePropertyType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Specific property type within a category
 */
export type PropertyResponsePropertyType =
  (typeof PropertyResponsePropertyType)[keyof typeof PropertyResponsePropertyType];

export const PropertyResponsePropertyType = {
  APARTMENT: 'APARTMENT',
  HOUSE: 'HOUSE',
  STUDIO: 'STUDIO',
  ROOM: 'ROOM',
  VILLA: 'VILLA',
  TOWNHOUSE: 'TOWNHOUSE',
  OTHER_RESIDENTIAL: 'OTHER_RESIDENTIAL',
  OFFICE: 'OFFICE',
  RETAIL: 'RETAIL',
  RESTAURANT: 'RESTAURANT',
  HOTEL: 'HOTEL',
  SHOWROOM: 'SHOWROOM',
  AUTO_DEALERSHIP: 'AUTO_DEALERSHIP',
  SNACKBAR: 'SNACKBAR',
  CAFE: 'CAFE',
  MOTEL: 'MOTEL',
  BAR: 'BAR',
  BED_AND_BREAKFAST: 'BED_AND_BREAKFAST',
  OTHER_COMMERCIAL: 'OTHER_COMMERCIAL',
  WAREHOUSE: 'WAREHOUSE',
  WORKSHOP: 'WORKSHOP',
  FACTORY: 'FACTORY',
  DATA_CENTER: 'DATA_CENTER',
  COLD_STORAGE: 'COLD_STORAGE',
  GARAGE: 'GARAGE',
  OTHER_INDUSTRIAL: 'OTHER_INDUSTRIAL',
  FARMLAND: 'FARMLAND',
  RANCH: 'RANCH',
  GREENHOUSE: 'GREENHOUSE',
  ORCHARD: 'ORCHARD',
  VINEYARD: 'VINEYARD',
  OTHER_AGRICULTURAL: 'OTHER_AGRICULTURAL',
  MIXED_USE: 'MIXED_USE',
  COMMERCIAL: 'COMMERCIAL',
} as const;