// app/src/generated/models/propertyResponseStatus.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Current occupancy or availability status of the property
 */
export type PropertyResponseStatus =
  (typeof PropertyResponseStatus)[keyof typeof PropertyResponseStatus];

export const PropertyResponseStatus = {
  VACANT: 'VACANT',
  OCCUPIED: 'OCCUPIED',
  SELF_OCCUPIED: 'SELF_OCCUPIED',
  MAINTENANCE: 'MAINTENANCE',
  UNAVAILABLE: 'UNAVAILABLE',
  UNDER_RENOVATION: 'UNDER_RENOVATION',
  FALLOW: 'FALLOW',
  LISTED: 'LISTED',
} as const;